+++ packages/cli/src/pages/api/previews/send.ts
@@ -2,37 +2,43 @@ import type { NextApiRequest, NextApiResponse } from "next";
 import { error } from "../../../util/log";
 import { sendMail } from "../../../moduleManifest";
 import { getPreviewComponent } from "../../../util/moduleManifestUtil";
+import { jsonStringifyError } from "../../../util/jsonStringifyError";
 
 export default async function send(
   req: NextApiRequest,
   res: NextApiResponse<SendPreviewResponseBody>
 ) {
   if (req.method !== "POST" || !req.body) {
-    res.writeHead(404);
-    res.end("Not found");
+    res.writeHead(405).end();
     return;
   }
-  const body: SendPreviewRequestBody = req.body;
 
+  const body: SendPreviewRequestBody = req.body;
   const { to, subject, previewClass, previewFunction } = body;
   let component;
+
   if (previewClass && previewFunction) {
     component = await getPreviewComponent(previewClass, previewFunction);
   }
-
   if (!component) {
     error("no component found");
-    res.writeHead(400);
-    res.end(JSON.stringify({ error: "no html provided, no component found" }));
+    res.status(400).json({ error: "no html provided, no component found" });
     return;
   }
 
-  await sendMail({
-    component,
-    to,
-    dangerouslyForceDeliver: true,
-    subject,
-  });
+  try {
+    await sendMail({
+      component,
+      to,
+      dangerouslyForceDeliver: true,
+      subject,
+    });
+    res.json({});
+  } catch (e: any) {
+    error("error sending mail", e);
 
-  res.json({});
+    res
+      .status(500)
+      .json({ error: "sendMail threw an error: " + jsonStringifyError(e) });
+  }
 }